rain rain

+++ 0x07-python-test_driven_development/2-matrix_divided.py
@@ -29,7 +29,7 @@ def matrix_divided(matrix, div):
             if type(val) is not int and type(val) is not float:
                 raise TypeError("matrix must be a matrix\
 (list of lists) of integers/floats")
-    if type(div) is not int and type(div) is not int:
+    if type(div) is not int and type(div) is not float:
         raise TypeError("div must be a number")
     if div == 0:
         raise ZeroDivisionError("division by zero")
@@ -38,4 +38,3 @@ def matrix_divided(matrix, div):
         for x in range(len(newlist[y])):
             newlist[y][x] = round(newlist[y][x] / div, 2)
     return (newlist)
-
